+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Starting from Rust port of the project, all changes will be put into this file.
 - `SJ`: Allow downloading "expired" chapters if you have subscription for it.
 - `SJ`: Fix failed deser on chapters
 - `SJ`: Show excerpt of data when deser fails
+- `SJ`: Don't show/download future chapter
 
 ### Build
 - Pin `windows-sys` dependencies (If you use crates.io, this is already pinned in `0.3.0`)